deleted old connection test moved shoes to lib/routes moved shoes.test to test/api
